+++ src/components/Header.js
@@ -1,0 +1,10 @@
+import * as Style from "components/PageStyle";
+
+export default function Header() {
+    return (
+        <Style.Header>
+            <h3>To Do List</h3>
+            <h4>React</h4>
+        </Style.Header>
+    );
+}
